Fixed errors in ByteArray read/write integers.

+++ lib/core/bytearrayprototype.h
@@ -34,29 +34,29 @@ public slots:
     double readDoubleBE(int offset = 0) const;
     float readFloatLE(int offset = 0) const;
     float readFloatBE(int offset = 0) const;
-    int32_t readInt32LE(int offset = 0) const;
-    int32_t readInt32BE(int offset = 0) const;
-    uint32_t readUInt32LE(int offset = 0) const;
-    uint32_t readUInt32BE(int offset = 0) const;
-    int16_t readInt16LE(int offset = 0) const;
-    int16_t readInt16BE(int offset = 0) const;
-    uint16_t readUInt16LE(int offset = 0) const;
-    uint16_t readUInt16BE(int offset = 0) const;
-    int8_t readInt8(int offset = 0) const;
+    int readInt32LE(int offset = 0) const;
+    int readInt32BE(int offset = 0) const;
+    uint readUInt32LE(int offset = 0) const;
+    uint readUInt32BE(int offset = 0) const;
+    int readInt16LE(int offset = 0) const;
+    int readInt16BE(int offset = 0) const;
+    uint readUInt16LE(int offset = 0) const;
+    uint readUInt16BE(int offset = 0) const;
+    int readInt8(int offset = 0) const;
 
     void writeDoubleLE(double v, int offset = 0);
     void writeDoubleBE(double v, int offset = 0);
     void writeFloatLE(float v, int offset = 0);
     void writeFloatBE(float v, int offset = 0);
-    void writeInt32LE(int32_t v, int offset = 0);
-    void writeInt32BE(int32_t v, int offset = 0);
-    void writeUInt32LE(uint32_t v, int offset = 0);
-    void writeUInt32BE(uint32_t v, int offset = 0);
-    void writeInt16LE(int16_t v, int offset = 0);
-    void writeInt16BE(int16_t v, int offset = 0);
-    void writeUInt16LE(uint16_t v, int offset = 0);
-    void writeUInt16BE(uint16_t v, int offset = 0);
-    void writeInt8(int8_t v, int offset = 0);
+    void writeInt32LE(int v, int offset = 0);
+    void writeInt32BE(int v, int offset = 0);
+    void writeUInt32LE(uint v, int offset = 0);
+    void writeUInt32BE(uint v, int offset = 0);
+    void writeInt16LE(int v, int offset = 0);
+    void writeInt16BE(int v, int offset = 0);
+    void writeUInt16LE(uint v, int offset = 0);
+    void writeUInt16BE(uint v, int offset = 0);
+    void writeInt8(int v, int offset = 0);
 
 private:
     QByteArray *thisByteArray() const;
